refactor: Update FollowButton type and enhance user list component
- Changed the `onChangeFollow` prop in `FollowButton` to accept a `React.MouseEvent`.
- Refactored `RecommendedUserList` to utilize new type definitions for user profile and paginated data.
- Improved cache management for follow/unfollow actions and updated user list rendering logic.
- Added `debouncedSearchQuery` prop to `RecommendedUserList` for better search handling.
