+++ src/app/(view)/follow/page.tsx
@@ -11,21 +11,10 @@ import { useRouter } from "next/navigation";
 import { useDebounce } from "@/_utils/useDebounce";
 import { getSearchUser } from "@/app/api/user/getSearchUser";
 import Image from "next/image";
-import { followUser } from "@/app/api/user/follow/followUser";
 import { IMyInfo } from "@/_types";
 import getMyInfo from "@/app/api/auth/getMyInfo";
-import { toast } from "react-toastify";
 import ServerToast from "@/_components/share/error/ServerToast";
 import SearchData from "@/_components/tasting-note/search/SearchData";
-import FollowButton from "@/_common/FollowButton";
-
-interface ISearchUser {
-  id: number;
-  isFollowed: boolean;
-  profileImage?: string;
-  nickname: string;
-  hasBadge: boolean;
-}
 
 const defaultProfileImage = `${process.env.NEXT_PUBLIC_IMAGE_BASE_PATH}/images/placeholders/profile/default_profile.png`;
 
@@ -50,7 +39,9 @@ export default function Page() {
 
   const router = useRouter();
   const [searchQuery, setSearchQuery] = useState<string>("");
-  const [searchQueryResult, setSearchQueryResult] = useState<ISearchUser[]>([]);
+  const [searchQueryResult, setSearchQueryResult] = useState<RecommendedUser[]>(
+    [],
+  );
   const debouncedSearchQuery = useDebounce(searchQuery, 250);
 
   const handleChangeQuery = useCallback(
@@ -63,30 +54,6 @@ export default function Page() {
   const handleClearSearchQuery = useCallback(() => {
     setSearchQuery("");
   }, []);
-
-  const handleClickUser = useCallback(
-    (id: number) => {
-      router.push(`/user/profile/${id}`);
-    },
-    [router],
-  );
-
-  const { mutate: handleFollow } = useMutation({
-    mutationFn: ({ id, isFollowed }: { id: number; isFollowed: boolean }) =>
-      followUser(id.toString()),
-    onSuccess: ({ id, nickname, isFollowed }) => {
-      setSearchQueryResult((prev) =>
-        prev.map((user) =>
-          user.id === id ? { ...user, isFollowed: !isFollowed } : user,
-        ),
-      );
-      toast(
-        isFollowed
-          ? `${nickname}님을 팔로우 취소했어요.`
-          : `${nickname}님을 팔로우했어요.`,
-      );
-    },
-  });
 
   useEffect(() => {
     const fetchData = async () => {
@@ -102,36 +69,12 @@ export default function Page() {
       } catch (error) {
         console.error("Failed to fetch search results:", error);
         setSearchQueryResult([]);
+        router.push("/");
       }
     };
 
     fetchData();
   }, [debouncedSearchQuery]);
-
-  const renderHighlightedName = useCallback(
-    (nickname: string) => {
-      if (!debouncedSearchQuery) return nickname;
-
-      const index = nickname
-        .toLowerCase()
-        .indexOf(debouncedSearchQuery.toLowerCase());
-      if (index === -1) return nickname;
-
-      return (
-        <>
-          {nickname.slice(0, index)}
-          <span className="text-primary-700">
-            {nickname.slice(index, index + debouncedSearchQuery.length)}
-          </span>
-          {nickname.slice(index + debouncedSearchQuery.length)}
-        </>
-      );
-    },
-    [debouncedSearchQuery],
-  );
-
-  const hasResults = debouncedSearchQuery && searchQueryResult.length > 0;
-  const noResults = debouncedSearchQuery && searchQueryResult.length === 0;
 
   if (isLoadingRecommendedSommelier) return <Loading />;
   if (error)
@@ -142,6 +85,7 @@ export default function Page() {
         cookieDelete
       />
     );
+
   // if (recommendedSommelier) {
   return (
     <div className="h-full w-full max-w-[560px]">
@@ -155,7 +99,7 @@ export default function Page() {
       />
       <div className="mb-4 h-[1px] w-full bg-cool-grayscale-300" />
 
-      {hasResults && (
+      {searchQuery.length > 0 && searchQueryResult.length > 0 && (
         <>
           <span className="mx-[4%] my-4 flex flex-row items-center text-sm text-cool-grayscale-500">
             <p className="text-cool-grayscale-800">
@@ -163,62 +107,15 @@ export default function Page() {
             </p>
             의 유저를 찾았어요.
           </span>
-          {searchQueryResult.map((user, index) => (
-            <div
-              key={user.id}
-              className={`border-t-[1px] border-cool-grayscale-200 ${
-                index === searchQueryResult.length - 1
-                  ? "border-b-[1px] border-cool-grayscale-200"
-                  : ""
-              }`}
-            >
-              <div className="mx-[4%] my-4 flex items-center justify-between">
-                <div
-                  className="flex cursor-pointer"
-                  onClick={() => handleClickUser(user.id)}
-                >
-                  <div>
-                    <Image
-                      width={44}
-                      height={44}
-                      src={user.profileImage ?? defaultProfileImage}
-                      alt="유저 이미지"
-                      className="rounded-full"
-                    />
-                  </div>
-                  <div className="ml-2 flex flex-row items-center gap-2">
-                    <p>{renderHighlightedName(user.nickname)}</p>
-                    {user.hasBadge && (
-                      <Image
-                        src={`${process.env.NEXT_PUBLIC_IMAGE_BASE_PATH}/images/kisa-badge.png`}
-                        alt="배지"
-                        className="mr-2 rounded-full"
-                        width={28}
-                        height={28}
-                      />
-                    )}
-                  </div>
-                </div>
-                <div>
-                  <FollowButton
-                    width="18"
-                    isFollowed={user.isFollowed}
-                    textSize="xs"
-                    onChangeFollow={() =>
-                      handleFollow({
-                        id: user.id,
-                        isFollowed: user.isFollowed,
-                      })
-                    }
-                  />
-                </div>
-              </div>
-            </div>
-          ))}
+          <RecommendedUserList
+            recommendedUserList={searchQueryResult}
+            userId={user!.memberId!.toString()}
+            debouncedSearchQuery={debouncedSearchQuery}
+          />
         </>
       )}
 
-      {noResults && (
+      {searchQuery.length > 0 && searchQueryResult.length === 0 && (
         <div className="flex min-h-screen w-full items-center justify-center">
           <div className="flex flex-col items-center justify-center">
             <p className="text-lg font-medium leading-7 text-cool-grayscale-600">
@@ -231,7 +128,7 @@ export default function Page() {
         </div>
       )}
 
-      {!hasResults && (
+      {searchQuery.length === 0  && (
         <>
           <div className="mx-[4%]">
             <p className="text-base font-medium leading-6 text-cool-grayscale-800">
