+++ src/app/api/user/getSearchUser.ts
@@ -7,22 +7,19 @@ export async function getSearchUser(params: {
   username: string;
 }) {
   const cookies = new Cookies();
-  try {
-    const headers = {
-      Authorization: `Bearer ${cookies.get("accessToken")}`,
-    };
-    const response = await axios.get(
-      `${process.env.NEXT_PUBLIC_JUULABEL_API_URL}/v1/api/members/search`,
-      {
-        params,
-        withCredentials: true,
-        headers,
-      },
-    );
-    if (response.status === 200 && response.data) return response.data;
-    return null;
-  } catch (error) {
-    console.error(error);
-    return null;
-  }
+
+  const headers = {
+    Authorization: `Bearer ${cookies.get("accessToken")}`,
+  };
+  const response = await axios.get(
+    `${process.env.NEXT_PUBLIC_JUULABEL_API_URL}/v1/api/members/search`,
+    {
+      params,
+      withCredentials: true,
+      headers,
+    },
+  );
+  if (response.status === 200 && response.data) return response.data;
+  console.log(response);
+  return null;
 }
